auth: drop the username/password login in participant app
fixes #841

Drop the `RegisterPage` component. There is no need for the AuthPageHOC
anymore, the HOC is converted to a simple component that is the
AuthPage, with the passwordless login form in place of the
WrappedComponent.

Remove the localLogin related routes from the participant router.
