+++ locales/en/auth.json
@@ -50,8 +50,6 @@
   "MagicLinkVerificationFailed": "Email link verification failed.",
   "MagicLinkUseAnotherMethod": "An error occurred while logging in. Please try again or use another authentication method",
   "CheckMagicEmail": "Connection information has been sent to the email address you requested. If you do not see it, make sure it is not in the spam.",
-  "UsePasswordLessAuthMethod": "Use authentication by email link",
-  "UseLocalLoginAuthMethod": "Use authentication with password",
   "UseNoLoginAuthMethod": "Currently, the only login method is through invitation. No other log in options are available at this time.",
   "WhyLoginInfo": "",
   "AnonymousLoginExplanation": "You will not be able to continue an incomplete survey if your session expires or if you change browsers.",
