+++ locales/fr/auth.json
@@ -50,8 +50,6 @@
   "MagicLinkVerificationFailed": "Échec de la vérification du lien courriel.",
   "MagicLinkUseAnotherMethod": "Une erreur est survenue au moment de l'authentification. Veuillez essayer à nouveau ou choisir une autre méthode d'authentification.",
   "CheckMagicEmail": "Les informations de connexion ont été envoyées à l'adresse courriel que vous avez entrée. Si vous ne les recevez pas, vérifiez vos pourriels.",
-  "UsePasswordLessAuthMethod": "Utiliser l'authentification par lien courriel",
-  "UseLocalLoginAuthMethod": "Utiliser l'authentification avec mot de passe",
   "UseNoLoginAuthMethod": "Actuellement, la seule méthode de connexion est via une invitation. Aucune autre option de connexion n'est disponible pour le moment.",
   "WhyLoginInfo": "",
   "AnonymousLoginExplanation": "Vous ne pourrez continuer une entrevue incomplète si votre session expire ou si vous changez de navigateur.",
