+++ packages/evolution-frontend/src/components/routers/ParticipantSurveyRouter.tsx
@@ -10,14 +10,11 @@ import { Route, Switch } from 'react-router-dom';
 import HomePage from '../pages/HomePage';
 import UnauthorizedPage from '../pages/UnauthorizedPage';
 import SurveyErrorPage from '../pages/SurveyErrorPage';
-import LoginPage from '../pages/auth/LoginPage';
-import RegisterWithPasswordPage from '../pages/auth/RegisterPage';
-import ForgotPasswordPage from 'chaire-lib-frontend/lib/components/pages/ForgotPasswordPage';
+import AuthPage from '../pages/auth/AuthPage';
 import MagicLinkVerifyPage from 'chaire-lib-frontend/lib/components/forms/auth/passwordless/MagicLinkVerify';
 import CheckMagicEmailPage from 'chaire-lib-frontend/lib/components/forms/auth/passwordless/CheckMagicEmail';
 import Survey from '../hoc/SurveyWithErrorBoundary';
 import PrivateRoute from 'chaire-lib-frontend/lib/components/routers/PrivateRoute';
-import ResetPasswordPage from 'chaire-lib-frontend/lib/components/pages/ResetPasswordPage';
 import PublicRoute from 'chaire-lib-frontend/lib/components/routers/PublicRoute';
 import ConsentedRoute from './ConsentedRoute';
 import config from 'chaire-lib-common/lib/config/shared/project.config';
@@ -34,10 +31,7 @@ const SurveyRouter = () => (
         <PublicRoute path="/" component={HomePage} exact={true} />
         <PublicRoute path={`${localesString}`} component={HomePage} exact={true} />
         <PublicRoute path="/home" component={HomePage} />
-        <ConsentedRoute path="/register" component={RegisterWithPasswordPage} />
-        <ConsentedRoute path="/login" component={LoginPage} />
-        <PublicRoute path="/forgot" component={ForgotPasswordPage} />
-        <PublicRoute path="/reset/:token" component={ResetPasswordPage} />
+        <ConsentedRoute path="/login" component={AuthPage} />
         <PublicRoute path="/unauthorized" component={UnauthorizedPage} />
         <PublicRoute path="/error" component={SurveyErrorPage} />
         <PublicRoute path="/magic/verify" component={MagicLinkVerifyPage} />
